+++ src/main.h
@@ -16,7 +16,7 @@ enum plotmode { PM_UNDEFINED, PM_NORMAL, PM_CONTINUATION, PM_RANGE, PM_PARTICLES
 /* function declaration */
 int odexp( oderhs pop_ode_rhs, oderhs single_rhs, odeic ode_ic, odeic single_ic, rootrhs root_rhs, const char *odexp_filename );
 
-int get_nbr_el(const char *filename, const char *sym, const size_t sym_len, size_t *nbr_el, size_t *nbr_epxr);
+int get_nbr_el(const char *filename, const char *sym, const int sym_len, int *nbr_el, int *nbr_epxr);
 
 int update_plot_options(int ngx, int ngy, int ngz, nve dxv);
 int update_plot_index(int *ngx, int *ngy, int *ngz, int *gx, int *gy, int *gz, nve dxv);
@@ -31,14 +31,14 @@ int save_snapshot(nve init, nve mu, double_array tspan, const char *odexp_filena
 
 int printf_options(const char *optiontype);
 
-void printf_list_val(char type, size_t print_index, size_t nve_index, int padding, const nve *var, char *descr);
-void printf_list_str(char type, size_t print_index, size_t nve_index, int padding, const nve *var);
+void printf_list_val(char type, int print_index, int nve_index, int padding, const nve *var, char *descr);
+void printf_list_str(char type, int print_index, int nve_index, int padding, const nve *var);
 
 void printf_SIM( void );
 
 int printf_status_bar( double_array *tspan );
 
-int gplot_data(const size_t colx, const size_t coly, const char *datafile_plotted );
+int gplot_data(const int colx, const int coly, const char *datafile_plotted );
 
 int gplot_particles( const int gx, const int gy, const nve var);
